+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 3.0.0 (unreleased)
 
-* `s3` – drop support for `aws-sdk-s3` versions lower than 1.14 (@janko)
+* `s3` – Remove `:host` option on `S3#initialize` (@janko)
 
-* `s3` – drop support for `aws-sdk` 2.x (@janko)
+* `s3` – Drop support for `aws-sdk-s3` versions lower than 1.14 (@janko)
+
+* `s3` – Drop support for `aws-sdk` 2.x (@janko)
 
 ## 2.19.0 (2019-07-18)